gdb: introduce xfer_partial_ctx object
Here's a patch I had sitting for a while.  I'm sending it to test the
water and collect comments.  Perhaps a subject of discussion for the GDB
BoF.

Between us and the the greater goal of reducing implicit references
to "current thing" global variables, such as current_inferior or
inferior_ptid stands the target_ops::xfer_partial interface.

The xfer_partial interface allows reading and writing all kinds of data
from the target, the most obvious one being memory.  The type of data
being read/written depends on the "object" passed to xfer_partial.
Certain objects implicitly require a given global context to be set.
This is not clearly documented, so we need to look at target
implementations to see what global state targets use when passed a given
object.  For example, when reading memory, most targets look at the
inferior_ptid variable.

This causes a ton of things in GDB (anything accessing memory, for
example) to require the current thread to be temporarily changed while
we do some operation.  I have always found this quite error-prone and
difficult to follow.

The xfer_partial interface also has an `annex` parameter, which is an
optional string.  Some objects require some data that refines the
request to be passed into `annex`.

The goal of what is proposed in this patch is to find a way to pass all
required context through the interface, instead of expecting targets to
access the global context.

An obvious quick hack would be to use the `annex` parameter for more.
For example, for memory requests, we could serialize a ptid in it and
deserialize it in the targets.  But that's really not ideal, especially
since we are free to change the interface as we see fit (given that we
update all users).

The proposal in this patch is to replace the `object` parameter of
`xfer_partial` with an object of type `xfer_partial_ctx`.
`xfer_partial_ctx` contains the object, but also any context associated
with that object.  The targets are therefore expected to read the
context from there instead of from global variable.

For example (and that's the only one I've really implemented as a PoC),
the memory objects need a ptid to read from, so its context needs a
ptid.  Another object might need an inferior, a pid, etc.  Using this
method, it becomes formalized in code what each object needs.

An `xfer_partial_ctx` object can only be created using one of the static
methods, so it becomes impossible to call `xfer_partial` to read memory
without passing a ptid.

In the end, the `annex` parameter would be eliminated, as whatever
passed in there would be passed in xfer_partial_ctx instead.

A possible shortcoming of this method is that different targets may need
different context for the same object.  For example, for a memory read,
a process target needs to know the ptid to read from.  The "exec"
target, which reads from an executable, does not read the ptid: it can
read memory even before there is a process.  So it accesses the current
inferior instead, which always exists.  We'll probably have to solve a
few problems like that.

Change-Id: I21cce403c5305e05c32d79058ea8056b86d3689c