Add to_value method to DWARF entry classes
Following the idea from the last patch this patch is adding another
conversion method from any DWARF entry object into a DWARF value
object.

Currently, we only know how to convert from a memory location
description into a value, but it is resonable to expect a set of target
hooks that would let the target decide on how to do other conversions
in the future.

gdb/ChangeLog:

        * dwarf2/expr.c (dwarf_entry::to_value): New method.
        (dwarf_location::to_value): New method.
        (dwarf_memory::to_value): New method.
        (ill_formed_expression): New function.

Change-Id: I5dfc7ed1a8c56970ce2ba5ce33f664d66d689628