Add to_location method to DWARF entry classes
DWARF standard already contains an implicit conversion between location
description and a DWARF value. In the DWARF 5 version, the only place
where this can happen is at the very end of the evaluation where a
client decides if the result is expected to be in a form of a value or
a location description (as_lval argument of the new evaluation method).

By allowing any location description to be on the DWARF stack, these
implicit conversions are expected on per operation basis which means
that the new DWARF entry classes need to have a support for those.

This patch adds a conversion method from any DWARF entry object into a
location description object.

gdb/ChangeLog:

        * dwarf2/expr.c (class dwarf_entry::to_location): New method.
        (class dwarf_value::to_location): New method.
        (class dwarf_location::to_location): New method.

Change-Id: Idbf10417f50f14191b6fe061cf83eb271edb8e8f