Add to_gdb_value method to DWARF entry class
The result of the DWARF expression evaluation is expected to be in a
format of a struct value object. This means that a new to_gdb_value
method is needed for all classes that derive from dwarf_entry class.

In the case of the dwarf_value class, it is usefull to have an even
more simplified version of this method (convert_to_gdb_value) that
does not require the evaluation frame context or any subobj
information. This method will be used even more in the following patch.

gdb/ChangeLog:

        * dwarf2/expr.c (dwarf_entry::to_gdb_value): New method.
        (dwarf_location::to_gdb_value): New method.
        (dwarf_value::convert_to_gdb_value): New method.
        (dwarf_value::to_gdb_value): New method.
        (dwarf_undefined::to_gdb_value): New method.
        (dwarf_memory::to_gdb_value): New method.
        (dwarf_register::to_gdb_value): New method.
        (dwarf_implicit::to_gdb_value): New method.
        (dwarf_implicit_pointer::to_gdb_value): New method.
        (dwarf_composite::to_gdb_value): New method.

Change-Id: I705ca1eb990ecaf0e2921317e9a082ab1f8a5643